requirements.txt
flask
flask-cors
werkzeug
psycopg2-binary
gunicorn
pandas